refactor(quartzite-style): extract Paint<LineEdit> into default_style/line_edit.rs
Verbatim lift of `impl Paint<LineEdit> for DefaultStyle` from mod.rs into a
new sibling module `default_style/line_edit.rs`, mirroring the existing
`text_edit` module structure. No behavioural change.
